refactor: created mock implementation
